код генерации биндингов разнесён
