+++ Assets/Code/Analytics/HandlersGeneration/SignalsBindingExtensions/SignalsBindingExtensionsGenerator.cs
@@ -1,10 +1,7 @@
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using Code.Analytics.GoogleSheetsIntegration;
-using Code.Extensions.Generation;
 
 namespace Code.Analytics.HandlersGeneration.SignalsBindingExtensions
 {
@@ -37,44 +34,8 @@ namespace Code.Analytics.HandlersGeneration.SignalsBindingExtensions
 				? string.Empty
 				: $"@this{NewLineWithSameOffset}{GenerateInvokes()};";
 
-		private string GenerateInvokes()
-		{
-			var result = new StringBuilder();
-			_handlers.ForEach((h) => result.Append(HandlerAsTemplate(h)));
-			return result.ToString();
-		}
+		private string GenerateInvokes() => string.Join(string.Empty, HandlersAsTemplates());
 
-		private string HandlerAsTemplate(AnalyticEventHandler handler)
-			=> BindingCodeTemplates.Method
-				(handler.Event, GetArgsDeclaration(handler.Parameters), GetArgsUsage(handler.Parameters));
-
-		private string GetArgsDeclaration(IEnumerable<(string type, string name)> parameters)
-			=> parameters.Any() ? ", v" : string.Empty;
-
-		private string GetArgsUsage(ICollection parameters)
-			=> parameters.Count switch
-			{
-				0 => string.Empty,
-				1 => "(v.Value)",
-				_ => GenerateArgsForTwoAndMore(parameters.Count)
-			};
-
-		private static string GenerateArgsForTwoAndMore(int parametersCount)
-		{
-			const int postfixLenght = 2;
-
-			var result = new StringBuilder();
-			result.Append('(');
-
-			for (var i = 0; i < parametersCount; i++)
-			{
-				result.Append($"v.Value{i + 1}, ");
-			}
-
-			result.RemoveLasSymbols(postfixLenght);
-			result.Append(')');
-
-			return result.ToString();
-		}
+		private IEnumerable<string> HandlersAsTemplates() => _handlers.Select(BindingCodeTemplates.Method);
 	}
 }
