+++ Assets/Code/Extensions/Generation/BindingGenerationArgumentsExtensions.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 67300f0752c24d8e84369099618a8088
+timeCreated: 1665934494
